BREAKING CHANGE: fix database model/entity (desc)
Database before this fix was generating empty entries, because floor library wasn't properly configured to utilize auto-generated primary key ids. This commit is marked as a breaking change, because it causes errors in LoadedTodosList, since model/entity api is using optional types now. However, this will be fixed in the next commit.