feat/refactor!: update presentation layer and implement adding todos (desc)
This commit adds the ability for the user to add tasks to the to-do list . Besides, it takes into account splitting the widget code into folders named according to the function the widgets perform. The names of folders and widgets may change in the future depending on the requirements of the project, and widgets can be divided into smaller ones according to the principle of single responsibility.